work on twitter button and linting
There is a twitter button that attempts to share the link to the specfic
program the user has run.

The linting function tries to be smarter and recognize whole numbers
better, as well as give more actionable feedback to the user.  I
implemented the concept of warnings for attempting to increment
unitialized memory locations